moving logs' state, mutations & input actions back to SPA
