+++ src/data/dataModule.js
@@ -1,54 +1,9 @@
 import {logFactory} from './logFactory';
 
 export default {
-  state: {
-    logs: [],
-    assets: [],
-    areas: [],
-    currentLogIndex: 0,
-  },
-
-  mutations: {
-    addLogs(state, logs) {
-      state.logs = state.logs.concat(logs);
-    },
-    addLogAndMakeCurrent(state, newLog) {
-      state.currentLogIndex = state.logs.push(newLog) -1;
-    },
-    updateCurrentLog (state, newProps) {
-      const updatedLog = logFactory({
-        ...state.logs[state.currentLogIndex],
-        ...newProps
-      });
-      state.logs.splice(state.currentLogIndex, 1, updatedLog);
-    },
-    clearLogs (state, payload) {
-      state.logs.splice(0, state.logs.length);
-    }
-  },
 
   actions: {
 
-    // INPUT ACTION
-    // TODO: Should this logic be moved to the 'addLogAndMakeCurrent' mutation?
-    //    Or perhaps just the logFactory, and pass in the date and logType as
-    //    a `newProps` object from a component method.
-    initializeLog({commit, rootState}, logType) {
-      // TODO: The User ID will also be needed to sync with server
-      const curDate = new Date(Date.now());
-      const timestamp = Math.floor(curDate / 1000).toString();
-      const curTimeString = curDate.toLocaleTimeString('en-US');
-      const curDateString = curDate.toLocaleDateString('en-US');
-      const newLog = logFactory({
-        type: logType,
-        name: `Observation: ${curDateString} - ${curTimeString}`,
-        field_farm_log_owner: rootState.user.name ? rootState.user.name : '',
-        timestamp: timestamp,
-      });
-      commit('addLogAndMakeCurrent', newLog);
-    },
-
-    // DB ACTION
     createRecord ({commit, dispatch, rootstate}, newRecord) {
       const tableName = newRecord.type
       delete newRecord.local_id
@@ -69,7 +24,6 @@ export default {
       })
     },
 
-    // DB ACTION
     loadCachedLogs({commit}, logType) {
       openDatabase()
       .then(function(db) {
@@ -89,10 +43,9 @@ export default {
       })
     },
 
-    // DB ACTION
     updateRecord ({commit, rootState}, newProps) {
       const newLog = logFactory({
-        ...rootState.data.logs[rootState.data.currentLogIndex],
+        ...rootState.farm.logs[rootState.farm.currentLogIndex],
         ...newProps
       });
       const table = newLog.type;
